+++ app/api/auth/[...nextauth]/auth.ts
@@ -1,0 +1,11 @@
+import GitHubProvider from "next-auth/providers/github";
+
+export const authOptions = {
+  providers: [
+    GitHubProvider({
+      clientId: process.env.GITHUB_ID as string,
+      clientSecret: process.env.GITHUB_SECRET as string,
+    }),
+  ],
+  secret: process.env.NEXTAUTH_SECRET,
+};
