+++ app/api/auth/[...nextauth]/route.ts
@@ -1,14 +1,5 @@
 import NextAuth from "next-auth";
-import GitHubProvider from "next-auth/providers/github";
-
-export const authOptions = {
-  providers: [
-    GitHubProvider({
-      clientId: process.env.GITHUB_ID as string,
-      clientSecret: process.env.GITHUB_SECRET as string,
-    }),
-  ],
-};
+import { authOptions } from "./auth";
 
 export const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
